requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
tqdm>=4.65.0
scikit-learn>=1.3.0

# Optional (for LLaVA)
sentencepiece>=0.1.99
protobuf>=3.20.0